Adicionado validation e logger

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -24,20 +24,46 @@ export class IdeaService {
 	}
 
 	async read(id: string) {
-		return await this.ideaRepository.findOne({
+		const idea = await this.ideaRepository.findOne({
 			where: { id },
 		});
+
+		if (!idea) {
+			throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+		}
+
+		return idea;
 	}
 
 	async update(id: string, data: Partial<IdeaDTO>) {
+		let idea = await this.ideaRepository.findOne({
+			where: { id },
+		});
+
+		if (!idea) {
+			throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+		}
+
 		await this.ideaRepository.update({ id }, data);
 
-		return await this.ideaRepository.findOne({ id });
+		idea = await this.ideaRepository.findOne({
+			where: { id },
+		});
+
+		return idea;
 	}
 
 	async destroy(id: string) {
+		const idea = await this.ideaRepository.findOne({
+			where: { id },
+		});
+
+		if (!idea) {
+			throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+		}
+
 		await this.ideaRepository.delete({ id });
 
-		return { deleted: true };
+		return idea;
 	}
 }
